+++ src/extension.ts
@@ -73,10 +73,10 @@ export async function activate(context: vscode.ExtensionContext) {
   which(config.get<string>('fortls.path'), (err: any) => {
     if (!config.get('ignoreWarning.fortls')) {
       if (err) {
-        const msg = `It is highly recommended to use the fortran-language-server to
+        const msg = `It is highly recommended to use the fortls to
               enable IDE features like hover, peeking, gotos and many more.
               For a full list of features the language server adds see:
-              https://github.com/hansec/fortran-language-server`;
+              https://github.com/gnikit/fortls`;
         promptForMissingTool(
           LANG_SERVER_TOOL_ID,
           msg,